Master Upload
This application was developed in Visual Basic for my Computing A level
in 2012. The application acts as a front end to a database created in
MySQL and hosted locally using WinAmp server. The app was designed for
the father of a friend who ran his own mechanics repair shop.
The app covers creation, storage and retrieval of the invoices,
customers and commonly used parts.
This was the first major application that I have created and as such, it
has some flaws. Chief amongst which is that it does not conform to data
protection laws and so I urge that no-one uses this system to store any
personal information.